+++ src/main/java/kitchenpos/exception/MenuProductsException.java
@@ -1,0 +1,8 @@
+package kitchenpos.exception;
+
+public class MenuProductsException extends RuntimeException {
+
+    public MenuProductsException(String message) {
+        super(message);
+    }
+}
